Refactor Colossus into Logger, placing billing metrics in Billing and send events in Event
