+++ src/index.ts
@@ -6,7 +6,8 @@ export * from './Router'
 export * from './VBase'
 export * from './ID'
 export * from './Workspaces'
-export * from './Colossus'
+export * from './Logger'
+export * from './Event'
 export * from './Metadata'
 export * from './responses'
 export * from './Billing'
